refactor: rename top rated data source files
