+++ data/src/main/java/com/london/data/local/database/NovixDatabase.kt
@@ -11,8 +11,8 @@ import com.london.data.local.database.convertor.SearchMoviesConverter
 import com.london.data.local.database.convertor.SearchTvShowConvertor
 import com.london.data.local.database.convertor.UpComingMovieTypeConverter
 import com.london.data.local.database.dao.home.popular.PopularSectionDao
+import com.london.data.local.database.dao.home.toprated.TopRatedDao
 import com.london.data.local.database.dao.home.upcoming.UpcomingSectionDao
-import com.london.data.local.database.dao.home.popular.TopRatedDao
 import com.london.data.local.database.dao.recent.search.RecentSearchDao
 import com.london.data.local.database.dao.recent.viewed.RecentViewedDao
 import com.london.data.local.database.dao.recent.whatched.movie.RecentWatchedMoviesDao
@@ -21,8 +21,8 @@ import com.london.data.local.database.dao.search.GenreInterestDao
 import com.london.data.local.database.dao.search.SearchActorsDao
 import com.london.data.local.database.dao.search.SearchMoviesDao
 import com.london.data.local.database.dao.search.SearchTvShowDao
-import com.london.data.local.model.home.popular.PopularSectionLocal
 import com.london.data.local.model.home.TopRatedLocal
+import com.london.data.local.model.home.popular.PopularSectionLocal
 import com.london.data.local.model.home.upcoming.UpComingSectionLocal
 import com.london.data.local.model.recent.search.RecentSearchLocal
 import com.london.data.local.model.recent.viewed.RecentViewedLocal
