+++ data/src/main/java/com/london/data/local/database/dao/home/HomeDao.kt
@@ -1,6 +1,6 @@
 @file:KoverIgnore
 
-package com.london.data.local.database.dao.home.popular
+package com.london.data.local.database.dao.home
 
 import com.london.domain.KoverIgnore
