+++ data/src/main/java/com/london/data/local/database/dao/home/toprated/TopRatedDao.kt
@@ -1,9 +1,10 @@
-package com.london.data.local.database.dao.home.popular
+package com.london.data.local.database.dao.home.toprated
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
+import com.london.data.local.database.dao.home.HomeDao
 import com.london.data.local.model.home.TopRatedLocal
 
 @Dao
